Graceful error if in .git or .hg file found
We're discovering the `.arcconfig` file by looking into a few
directories, one of which is `./.git/` or `./.hg/`. If in current
directory a `./.git` or `./.hg` file has been created, an `IOError` `ENOTDIR`
is raised. The `./.git` file is part of any git worktree and the user
might accidently run moz-phab from such directory.

This patch makes the code behave in the same way as the file wouldn't
exist (`ENOENT`) and fail with the "Failed to determine Phabricator URL
(missing .arcconfig?)" message.